+++ data/onboarding/src/main/java/com/hilingual/data/onboarding/di/OnboardingDataStoreModule.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.hilingual.core.localstorage.builder.DataStoreBuilder
+import com.hilingual.data.onboarding.di.qualifier.OnboardingDataStore
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
