+++ data/onboarding/src/main/java/com/hilingual/data/onboarding/di/qualifier/OnboardingDataStore.kt
@@ -1,4 +1,4 @@
-package com.hilingual.data.onboarding.di
+package com.hilingual.data.onboarding.di.qualifier
 
 import javax.inject.Qualifier
